+++ internal/services/location_service_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func setupLocationService(t *testing.T) (services.LocationService, func()) {
+func setupLocationService(t *testing.T) (services.LocationService, *services.MarkerService, func()) {
 	t.Helper()
 	dbc, cleanup := setupDB(t)
 
@@ -21,11 +21,11 @@ func setupLocationService(t *testing.T) (services.LocationService, func()) {
 	blockRepo := repositories.NewBlockRepository(dbc, blockStateRepo)
 	markerService := services.NewMarkerService(markerRepo)
 	locationService := services.NewLocationService(clueRepo, locationRepo, markerRepo, blockRepo, markerService)
-	return locationService, cleanup
+	return locationService, markerService, cleanup
 }
 
 func TestLocationService_CreateLocation(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, _, cleanup := setupLocationService(t)
 	defer cleanup()
 
 	t.Run("Create location", func(t *testing.T) {
@@ -64,10 +64,10 @@ func TestLocationService_CreateLocation(t *testing.T) {
 }
 
 func TestLocationService_CreateLocationFromMarker(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, markerService, cleanup := setupLocationService(t)
 	defer cleanup()
 
-	marker, err := service.CreateMarker(
+	marker, err := markerService.CreateMarker(
 		context.Background(),
 		gofakeit.Name(),
 		gofakeit.Latitude(),
@@ -117,7 +117,7 @@ func TestLocationService_CreateLocationFromMarker(t *testing.T) {
 }
 
 func TestLocationService_DuplicateLocation(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, _, cleanup := setupLocationService(t)
 	defer cleanup()
 
 	blockService, blockCleanup := setupBlocksService(t)
@@ -156,7 +156,7 @@ func TestLocationService_DuplicateLocation(t *testing.T) {
 }
 
 func TestLocationService_GetByID(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, _, cleanup := setupLocationService(t)
 	defer cleanup()
 
 	t.Run("Get location by ID", func(t *testing.T) {
@@ -181,7 +181,7 @@ func TestLocationService_GetByID(t *testing.T) {
 }
 
 func TestLocationService_GetByInstanceAndCode(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, _, cleanup := setupLocationService(t)
 	defer cleanup()
 
 	t.Run("Get location by instance and code", func(t *testing.T) {
@@ -206,7 +206,7 @@ func TestLocationService_GetByInstanceAndCode(t *testing.T) {
 }
 
 func TestLocationService_FindByInstance(t *testing.T) {
-	service, cleanup := setupLocationService(t)
+	service, _, cleanup := setupLocationService(t)
 	defer cleanup()
 
 	t.Run("Find locations by instance", func(t *testing.T) {